requirements.txt
fastapi==0.110.0
uvicorn==0.27.1

# AWS only if needed later
boto3==1.34.90

# ML stack (CPU wheels)
torch==2.1.2
sentence-transformers==2.6.1
transformers==4.39.3

# Vector search
faiss-cpu==1.7.4

# Utils
numpy<2.0
pydantic==2.6.4
